feat: add WalletMapper for entity to DTO conversions

+++ src/main/java/com/example/walletrest/infrastructure/entity/Asset.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
@@ -20,10 +22,10 @@ public class Asset {
     private String symbol;
 
     @Column(nullable = false)
-    private double quantity;
+    private BigDecimal quantity;
 
     @Column(nullable = false)
-    private double price;
+    private BigDecimal price;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "wallet_id")
